pyproject.toml
[tool.poetry]
name = "anki_assist_ai"
version = "0.1.0"
description = "A project to create Anki cards using AnkiConnect"
authors = ["Ehsan Olyaee"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
openai = "^1.53.0"
pillow = "^10.0.0"
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"
streamlit = "^1.39.0"
toml = "^0.10.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "C90"]
ignore = ["E501"]